day-01 part-01

+++ main/__TEMPLATE__/solution.cpp
@@ -2,7 +2,7 @@
 
 #include <istream>
 
-int solution_part_one(const std::vector<std::string> &input) {
+long solution_part_one(const std::vector<std::string> &input) {
     for (const auto &line: input) {
         std::print("{}\n", line);
     }
@@ -10,7 +10,7 @@ int solution_part_one(const std::vector<std::string> &input) {
     return 0;
 }
 
-int solution_part_two(const std::vector<std::string> &input) {
+long solution_part_two(const std::vector<std::string> &input) {
     for (const auto &line: input) {
         std::print("{}\n", line);
     }
