+++ main/__TEMPLATE__/solution.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <vector>
 
-int solution_part_one(const std::vector<std::string> &);
+long solution_part_one(const std::vector<std::string> &);
 
-int solution_part_two(const std::vector<std::string> &);
+long solution_part_two(const std::vector<std::string> &);
